Align the tests.
